functions/requirements.txt
google-cloud-storage==2.10.0
Pillow==10.0.0
Flask==2.1.3
Werkzeug==2.0.3
functions-framework==3.4.0
